+++ src/resourceManager/resourceManager.h
@@ -16,11 +16,11 @@ public:
 
     static ResourceManager& instance();
 
-	void loadModel(int entityID, std::string const& path);
+	void loadModel(size_t entityID, std::string const& path);
 
-	std::vector<Mesh>* getMeshes(const int entityID) { return &mEntityMeshes.at(entityID); }
+	std::vector<Mesh>* getMeshes(const size_t entityID) { return &mEntityMeshes.at(entityID); }
 
-	std::vector<Texture>* getTextures(const int entityID) { return &mEntityTextures.at(entityID); }
+	std::vector<Texture>* getTextures(const size_t entityID) { return &mEntityTextures.at(entityID); }
 
 private:
 	explicit ResourceManager() = default;
@@ -36,8 +36,8 @@ private:
 	std::string mDirectory;
 	std::vector<Mesh> mMeshes;
 	std::vector<Texture> mTextures;
-	std::unordered_map<int, std::vector<Mesh>> mEntityMeshes;
-	std::unordered_map<int, std::vector<Texture>> mEntityTextures;
+	std::unordered_map<size_t, std::vector<Mesh>> mEntityMeshes;
+	std::unordered_map<size_t, std::vector<Texture>> mEntityTextures;
 	std::unordered_set<std::string> mTexturesLoaded;
 };
